requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
pandas
plotly
python-dotenv